Add files via upload
I implemented a menu and two upgrades.  The menu currently works.  I cannot provide screenshots because I am currently in the middle of fixing an error (again).  The app currently looks exactly the same as it did in class on Thursday.  When you click the menu button, you get a menu with 2 buttons and label that says how many coins you have.  The first upgrade makes it so that every time you click, you get 2 coins instead of one.  This upgrade will cost 100 coins, but currently costs 10 because I need to be able to easily test everything.  You cannot buy the upgrade until you have enough coins for it and it vanishes once you've bought it (so you can't buy it again).  The second upgrade will give you one free coin every 10 seconds and will cost a large amount of coins to buy (maybe 250 or 500), but currently costs nothing and gives you one free coin every second because I need to be able to test it.  This upgrade button is also supposed to vanish once you've bought it.  After you've bought